firewall.client: Allow to remove ipv:address pair for service destinations II
This makes removeDestination compatible to other remove methods.

Now also in FirewallClientConfigService

+++ src/firewall/client.py
@@ -1402,7 +1402,10 @@ class FirewallClientConfigService(object):
 
     @slip.dbus.polkit.enable_proxy
     @handle_exceptions
-    def removeDestination(self, destination):
+    def removeDestination(self, destination, address=None):
+        if address is not None and self.getDestination(destination) != address:
+            raise FirewallError(errors.NOT_ENABLED, "'%s:%s'" % \
+                                (destination, address))
         self.fw_service.removeDestination(destination)
 
     @slip.dbus.polkit.enable_proxy
